chore: update tsconfig and rename test utilities
