+++ package.json
@@ -75,7 +75,6 @@
     "@storybook/addon-onboarding": "^1.0.11",
     "@storybook/addon-toolbars": "^7.6.17",
     "@storybook/addon-viewport": "^7.6.17",
-    "@storybook/addon-webpack5-compiler-swc": "^7.6.17",
     "@storybook/blocks": "^7.6.17",
     "@storybook/builder-webpack5": "^7.6.17",
     "@storybook/react": "^7.6.17",
